Projetos/Admin/src/java/web/java/classe/CursoBean.java
package web.java.classe;

/**
 *
 * @author Marcio Augusto Schlosser
 */
public class CursoBean {
    private int id, idInstituicao;
    private String nome;

    public int getId() {
        return id;
    }

    public void setId(int id) {
        this.id = id;
    }

    public int getIdInstituicao() {
        return idInstituicao;
    }

    public void setIdInstituicao(int idInstituicao) {
        this.idInstituicao = idInstituicao;
    }
    
    public String getNome() {
        return nome;
    }

    public void setNome(String nome) {
        this.nome = nome;
    }
}
